requirements.txt
langchain
langchain-community
langchain-core
python-dotenv
langsmith
langgraph
ollama
tqdm
jpype1
